chore(#issue-17): codereview, separate logic to clean up implementation/test
